# backend-api/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Weather API",
    "description": "JSON REST API for viewing weather data"
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/users/login": {
      "post": {
        "summary": "Logs in the user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Logs in the user",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "password": "string"
              },
              "example": {
                "email": "example@email.com",
                "password": "password"
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "summary": "Logs out the user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Logs out the user",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "88f76721-ff89-46f2-ac3a-c0c8a24d5625"
              },
              "example": {
                "authenticationKey": "88f76721-ff89-46f2-ac3a-c0c8a24d5625"
              }
            }
          }
        }
      }
    },
    "/users/{authenticationKey}": {
      "get": {
        "summary": "Get all users",
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{id}/{authenticationKey}": {
      "get": {
        "summary": "Get a user by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authenticationKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Gets a user by id"
        }
      }
    },
    "/users/get-by/{authenticationKey}": {
      "get": {
        "summary": "Get a user by authentication key",
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Gets a user by their authenticaion key"
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Create a user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Creates a user",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "password": "string",
                "role": "string",
                "firstName": "string",
                "lastName": "string,",
                "authenticationKey": "string"
              },
              "example": {
                "email": "student@email.com",
                "password": "abc123",
                "role": "student",
                "firstName": "test",
                "lastName": "test",
                "authenticationKey": "authKey"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a user by id",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Updates a user by id",
          "content": {
            "application/json": {
              "schema": {
                "id": "string",
                "email": "string",
                "password": "string",
                "role": "string",
                "firstName": "string",
                "lastName": "string,",
                "authenticationKey": "string"
              },
              "example": {
                "id": "640ea49386348c3384bb4419",
                "email": "student@email.com",
                "password": "abc123",
                "role": "student",
                "firstName": "test",
                "lastName": "test",
                "authenticationKey": "authKey"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete users that haven't logged in for more than 30 days",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Authenticaion key required to delete many",
          "content": {
            "application/json": {
              "schema": {},
              "example": {
                "authenticationKey": "88f76721-ff89-46f2-ac3a-c0c8a24d5625"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "summary": "Delete a user by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Deletes a single user",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "string"
              },
              "example": {
                "authenticationKey": "authKey"
              }
            }
          }
        }
      }
    },
    "/usersLastLogin": {
      "get": {
        "summary": "Get all users from last login first",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/deleteUsers": {
      "delete": {
        "summary": "Delete many users by id",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Deletes many users by id",
          "content": {
            "application/json": {
              "schema": {
                "ids": [
                  "string"
                ],
                "authenticationKey": "string"
              },
              "example": {
                "authenticationKey": "0ec6cd91-ad1c-4c60-9e51-cc1b32a11f6a",
                "ids": [
                  "640ea49386348c3384bb4419",
                  "640ea49386348c3384bb4419"
                ]
              }
            }
          }
        }
      }
    },
    "/updateAccessLevel": {
      "patch": {
        "summary": "Update access level for at least two users in the same query, based on a date range in which the users were created",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "description": "Updates access level for at least two users in the same query, based on a date range in which the users were created",
          "content": {
            "application/json": {
              "schema": {
                "role": "string",
                "authenticationKey": "string",
                "startDate": "string",
                "endDate": "string"
              },
              "example": {
                "role": "admin",
                "authenticationKey": "4f7ecabb-dbc8-4f9f-9749-f4cfba6fa68e",
                "startDate": "27/01/2023",
                "endDate": "29/01/2023"
              }
            }
          }
        }
      }
    },
...
  }
}